+++ spec/resources/microsoft_remote_desktop_app/mac_os_x/10_10_spec.rb
@@ -7,12 +7,13 @@ describe 'resource_microsoft_remote_desktop_app::mac_os_x::10_10' do
                              platform: 'mac_os_x',
                              version: '10.10')
   end
-  let(:chef_run) do
+  let(:converge) do
     runner.converge("microsoft_remote_desktop_app_test::#{action}")
   end
 
   context 'the default action (:install)' do
     let(:action) { :default }
+    cached(:chef_run) { converge }
 
     it 'configures the Mac App Store' do
       expect(chef_run).to include_recipe 'mac-app-store'